remove state and move metadata
