+++ src/pymmcore_plus/mda/__init__.py
@@ -3,16 +3,11 @@ from ._protocol import PMDAEngine
 from ._runner import MDARunner
 from ._thread_relay import mda_listeners_connected
 from .events import PMDASignaler
-from .metadata import FrameMetaV1, SummaryMetaV1, frame_metadata, summary_metadata
 
 __all__ = [
-    "frame_metadata",
-    "FrameMetaV1",
     "mda_listeners_connected",
     "MDAEngine",
     "MDARunner",
     "PMDAEngine",
     "PMDASignaler",
-    "summary_metadata",
-    "SummaryMetaV1",
 ]
