+++ src/pymmcore_plus/mda/handlers/_ome_tiff_writer.py
@@ -48,7 +48,7 @@ if TYPE_CHECKING:
 
     import useq
 
-    from pymmcore_plus.mda.metadata import SummaryMetaV1
+    from pymmcore_plus.metadata import SummaryMetaV1
 
 IMAGEJ_AXIS_ORDER = "tzcyxs"
