+++ src/pymmcore_plus/metadata/__init__.py
