+++ src/pymmcore_plus/metadata/serialize.py
