+++ tests/test_core.py
@@ -611,33 +611,10 @@ def test_set_autofocus_offset(
 
 
 def test_core_state(core: CMMCorePlus) -> None:
-    state = core.state(
-        devices=True,
-        image=True,
-        system_info=True,
-        system_status=True,
-        config_groups=True,
-        position=True,
-        autofocus=True,
-        pixel_size_configs=True,
-        device_types=True,
-    )
+    state = core.state()
     assert isinstance(state, dict)
-    assert "Devices" in state
-    assert "Core" in state["Devices"]
-
-    for key in {
-        "AutoFocus",
-        "Camera",
-        "Focus",
-        "Galvo",
-        "ImageProcessor",
-        "SLM",
-        "Shutter",
-        "XYStage",
-    }:
-        if val := state["Devices"]["Core"][key]:
-            assert val in state["Devices"]
+    assert "devices" in state
+    assert any(d["name"] == "Core" for d in state["devices"])
 
     core.unloadAllDevices()
     # should still work without error
